fix(cryptography): output raw 16 byte char array for MD5
The RFC spec dictates that the output of the MD5 function should be the raw 16 byte digest. The
previous implementation was not following this spec and was instead outputting the 32 byte string
representation of the hex digest. This update ensure the MD5 function operate as per the RFC spec.

BREAKING CHANGE: Any usages of the MD5 function will now output a raw 16 byte digest as per the RFC
spec. If you need the 32 byte string representation of the digest the output of the MD5 function
will now need to converted manually. The Encoding library provides such a function, called
NAVHexToString that can be used.